Update ServiceDialog to handle provider edits to quantity. Move request fetch from RequestMetadata to RequestDetails.
