+++ src/components/RequestMetadata.tsx
@@ -1,71 +1,6 @@
-import React, { useEffect } from "react";
-import { useAppSelector, useAppDispatch } from "@/app/hooks";
-import { setRequest } from "@/features/requests/requestDetailsSlice";
-import { useNavigate, useParams } from "react-router-dom";
-import createApi from "@/lib/api";
-import { useToast } from "@/hooks/use-toast";
+import React from "react";
 
-function RequestMetadata() {
-    const dispatch = useAppDispatch();
-    const navigate = useNavigate();
-    const { toast } = useToast();
-    const { requestId } = useParams();
-    const request = useAppSelector((state) => state.request.request);
-
-    useEffect(() => {
-        const controller = new AbortController();
-        const api = createApi({ endpoint: `/requests/${requestId}` })
-
-        async function getData() {
-            try {
-                const response = await api.get({ controller: controller });
-                const json = await response.json();
-
-                if (!response.ok) {
-                    if (response.status == 401) {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                        navigate("/login");
-                    } else if (response.status == 403) {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                        navigate("/");
-                    } else if (response.status == 404) {
-                        toast({
-                            variant: "destructive",
-                            description: "Request does not exist"
-                        });
-                        navigate("/");
-                    } else {
-                        toast({
-                            variant: "destructive",
-                            description: `${json.errors.detail}`
-                        });
-                    }
-                }
-                else {
-                    dispatch(setRequest({ ...json.data }));
-                    console.log(json.data)
-                }
-            } catch (error) {
-                toast({
-                    variant: "destructive",
-                    description: `${error}`,
-                });
-            }
-        }
-
-        getData();
-
-        return () => {
-            controller.abort("Request Aborted");
-        }
-    }, []);
-
+function RequestMetadata({ request }) {
     return <>
         <dl className="flex flex-col md:flex-row p-3">
             <dt className="w-32 font-semibold">Coordinator</dt>
